backend/requirements.txt
certifi
click
cryptography
Flask
Flask-Cors
Flask-SQLAlchemy
greenlet
gunicorn
idna
Jinja2
limits
MarkupSafe
PyMySQL
python-dotenv
requests
six
SQLAlchemy
urllib3
Werkzeug